Logout, delete confirmation, image pic and upload
Got logout to fully work, got the delete to ask if the user is sure as well as fixed the bug on the last question not being removed live. i also worked on taking images on a phone and i got the pictures taken and able to pull from albums folder.